+++ tests/testthat/test-totals.R
@@ -12,8 +12,9 @@ exp <- tibble::tribble(
   "045.0",  1000,
 )
 
-disc <- calc_discount(wp, TRUE, NA, FALSE)
+disc <- calc_discount(wp, TRUE, 3, FALSE)
 disc2 <- calc_discount(wp, FALSE, 5, FALSE)
+disc3 <- calc_discount(wp, FALSE, 0, FALSE)
 
 fte <- list(fte = FALSE)
 fte2 <- list(fte = TRUE,
@@ -23,6 +24,7 @@ fte2 <- list(fte = TRUE,
              ))
 
 oh <- overhead(disc)
+oh3 <- overhead(disc3)
 
 test_that("internal totals", {
   tot <- totals(wp,
@@ -34,7 +36,7 @@ test_that("internal totals", {
                 fte = fte)
 
   expect_equal(nrow(tot), 5)
-  expect_equal(unlist(tot[grepl("number of hours", tot$Description),2]),
+  expect_equal(unlist(tot[grepl("Discount \\(3%\\)", tot$Description),2]),
                "  -90", ignore_attr = TRUE)
   expect_equal(unlist(tot[grepl("Total", tot$Description),2]), "5'301", ignore_attr = TRUE)
   expect_true("Operating costs (10%)" %in% tot$Description)
@@ -48,10 +50,21 @@ test_that("internal totals", {
                  snf = FALSE,
                  fte = fte)
   expect_equal(nrow(tot2), 5)
-  expect_equal(unlist(tot2[grepl("number of hours", tot$Description),2]), " -150", ignore_attr = TRUE)
+  expect_equal(unlist(tot2[grepl("Discount \\(5%\\)", tot2$Description),2]), " -150", ignore_attr = TRUE)
   expect_equal(unlist(tot2[grepl("Total", tot2$Description),2]), "5'241", ignore_attr = TRUE)
   expect_true("Operating costs (10%)" %in% tot2$Description)
   expect_true(!"University overhead (10%)" %in% tot2$Description)
+
+
+  tot3 <- totals(wp,
+                 expenses = exp,
+                 discount = disc3,
+                 discount_chf = 1000,
+                 overhead = oh3,
+                 rate = "External non-profit",
+                 snf = FALSE,
+                 fte = fte)
+  expect_equal(unlist(tot3[grepl("Discount", tot3$Description),2]), "-1'000", ignore_attr = TRUE)
 })
 
 test_that("external totals", {
@@ -63,7 +76,7 @@ test_that("external totals", {
                 snf = FALSE,
                 fte = fte)
   expect_equal(nrow(tot3), 6)
-  expect_equal(unlist(tot3[grepl("number of hours", tot3$Description),2]), "  -90", ignore_attr = TRUE)
+  expect_equal(unlist(tot3[grepl("Discount \\(3%\\)", tot3$Description),2]), "  -90", ignore_attr = TRUE)
   expect_equal(unlist(tot3[grepl("Total", tot3$Description),2]), "5'692", ignore_attr = TRUE)
   expect_equal(unlist(tot3[grepl("University overhead", tot3$Description),2]),
                unlist(tot3[grepl("Operating costs", tot3$Description),2]), ignore_attr = TRUE)
@@ -79,7 +92,7 @@ test_that("external totals", {
                  snf = FALSE,
                  fte = fte)
   expect_equal(nrow(tot4), 6)
-  expect_equal(unlist(tot4[grepl("number of hours", tot4$Description),2]), " -150", ignore_attr = TRUE)
+  expect_equal(unlist(tot4[grepl("Discount \\(5%\\)", tot4$Description),2]), " -150", ignore_attr = TRUE)
   expect_equal(unlist(tot4[grepl("Total", tot4$Description),2]), "5'632", ignore_attr = TRUE)
   expect_equal(unlist(tot4[grepl("University overhead", tot4$Description),2]), unlist(tot4[3,2]), ignore_attr = TRUE)
   expect_true("Operating costs (10%)" %in% tot4$Description)
@@ -97,7 +110,7 @@ test_that("fte totals", {
                 snf = FALSE,
                 fte = fte2)
   expect_equal(nrow(tot3), 7)
-  expect_equal(unlist(tot3[grepl("number of hours", tot3$Description),2]), "  -90", ignore_attr = TRUE)
+  expect_equal(unlist(tot3[grepl("Discount \\(3%\\)", tot3$Description),2]), "  -90", ignore_attr = TRUE)
   expect_equal(unlist(tot3[grepl("Total", tot3$Description),2]), "5'792", ignore_attr = TRUE)
   expect_equal(unlist(tot3[grepl("University overhead", tot3$Description),2]),
                unlist(tot3[grepl("Operating costs", tot3$Description),2]), ignore_attr = TRUE)
@@ -114,7 +127,7 @@ test_that("fte totals", {
                  snf = FALSE,
                  fte = fte2)
   expect_equal(nrow(tot4), 6)
-  expect_equal(unlist(tot4[grepl("number of hours", tot4$Description),2]), " -150", ignore_attr = TRUE)
+  expect_equal(unlist(tot4[grepl("Discount \\(5%\\)", tot4$Description),2]), " -150", ignore_attr = TRUE)
   expect_equal(unlist(tot4[grepl("Total", tot4$Description),2]), "5'341", ignore_attr = TRUE)
   expect_equal(unlist(tot4[grepl("FTE", tot4$Description),2]), "  100", ignore_attr = TRUE)
   expect_true("Operating costs (10%)" %in% tot4$Description)